+++ src/Mailer.php
@@ -12,7 +12,6 @@ use Psr\Log\LoggerInterface;
 use Symfony\Component\Mailer\Transport;
 use Symfony\Component\Mailer\Transport\Dsn;
 use Symfony\Component\Mailer\Transport\TransportInterface;
-use Yii;
 use yii\base\InvalidArgumentException;
 use yii\base\InvalidConfigException;
 use yii\mail\BaseMailer;
@@ -20,7 +19,9 @@ use yii\psr\DynamicLogger;
 
 /**
  * @psalm-suppress PropertyNotSetInConstructor
- * @psalm-type PsalmTransportConfig array{scheme?:string, host?:string, username?:string, password?:string, port?:int, options?: array, dsn?:string|Dsn }
+ * @psalm-type TransportHostArray array{scheme?:string, host?:string, username?:string, password?:string, port?:int, options?: array<mixed>, dsn?:string|Dsn }
+ * @phpstan-type TransportConfigArray array{scheme?:string, host?:string, username?:string, password?:string, port?:int, options?: array<mixed>, dsn?:string|Dsn }
+ * @extendable
  */
 class Mailer extends BaseMailer
 {
@@ -37,6 +38,9 @@ class Mailer extends BaseMailer
      * @see https://symfony.com/doc/current/mailer.html#signing-messages
      */
     public ?MessageSignerInterface $signer = null;
+    /**
+     * @var array<mixed>
+     */
     public array $signerOptions = [];
     /**
      * @var null|TransportInterface Symfony transport instance or its array configuration.
@@ -45,15 +49,11 @@ class Mailer extends BaseMailer
     public ?Transport $transportFactory = null;
 
     /**
-     * @param PsalmTransportConfig|TransportInterface $transport
+     * @param TransportConfigArray|TransportInterface $transport
      * @throws InvalidConfigException on invalid argument.
      */
-    public function setTransport($transport): void
+    public function setTransport(array|TransportInterface $transport): void
     {
-        if (!is_array($transport) && !$transport instanceof TransportInterface) {
-            throw new InvalidArgumentException('"' . get_class($this) . '::transport" should be either object or array, "' . gettype($transport) . '" given.');
-        }
-
         $this->_transport = $transport instanceof TransportInterface ? $transport : $this->createTransport($transport);
     }
 
@@ -73,25 +73,24 @@ class Mailer extends BaseMailer
         }
         /** @var LoggerInterface|null $logger */
         $logger = class_exists(DynamicLogger::class) ? new DynamicLogger() : null;
-        /**
-         * @psalm-suppress TooManyArguments On PHP 7.4 symfony/mailer 5.4 is uses which uses func_get_args instead of real args
-         */
         $defaultFactories = Transport::getDefaultFactories(new EventDispatcherProxy($this), null, $logger);
         /** @psalm-suppress InvalidArgument Symfony's type annotation is wrong */
         return new Transport($defaultFactories);
     }
 
     /**
-     * @param PsalmTransportConfig $config
+     * @param TransportConfigArray $config
      * @throws InvalidConfigException
      */
     private function createTransport(array $config = []): TransportInterface
     {
         $transportFactory = $this->getTransportFactory();
-        if (array_key_exists('dsn', $config) && is_string($config['dsn'])) {
-            $transport = $transportFactory->fromString($config['dsn']);
-        } elseif (array_key_exists('dsn', $config) && $config['dsn'] instanceof Dsn) {
-            $transport = $transportFactory->fromDsnObject($config['dsn']);
+        if (array_key_exists('dsn', $config)) {
+            if (is_string($config['dsn'])) {
+                $transport = $transportFactory->fromString($config['dsn']);
+            } else {
+                $transport = $transportFactory->fromDsnObject($config['dsn']);
+            }
         } elseif (array_key_exists('scheme', $config) && array_key_exists('host', $config)) {
             $dsn = new Dsn(
                 $config['scheme'],